extract World
