+++ src/test/java/ru/mercuriev/game/of/life/no/condition/NeighboursCollectorTest.java
@@ -14,7 +14,8 @@ public class NeighboursCollectorTest implements Fixture {
 
     @Test
     public void testCollect() throws Exception {
-        WrappedWorld wrappedWorld = WrappedWorld.newInstance(GENERATION);
+        WrappedWorld build = WrappedWorld.newInstance(GENERATION);
+        WrappedWorld wrappedWorld = WrappedWorld.newInstance(build);
         NeighboursCollector collector = new NeighboursCollector(wrappedWorld);
         List<Cell> neighbours = collector.collect(wrappedWorld.cellAt(2, 0));
         assertNotNull(neighbours);
@@ -23,7 +24,8 @@ public class NeighboursCollectorTest implements Fixture {
 
     @Test
     public void testCollectFull() {
-        WrappedWorld w = WrappedWorld.newInstance(GENERATION);
+        WrappedWorld build = WrappedWorld.newInstance(GENERATION);
+        WrappedWorld w = WrappedWorld.newInstance(build);
         NeighboursCollector c = new NeighboursCollector(w);
         IntStream.range(0, w.size()).forEach(i ->
                 IntStream.range(0, w.size()).forEach(j -> {
